[hw8-architecture-patterns] (Refactoring; Bug fixes)

+++ email-service/internal/module/mail/migrations/000002_create_table_currencies.up.sql
@@ -3,6 +3,6 @@ CREATE SEQUENCE currencies_id_seq;
 CREATE TABLE currencies
 (
     id     BIGINT DEFAULT nextval('currencies_id_seq') NOT NULL PRIMARY KEY,
-    number BIGINT                                      NOT NULL,
+    number FLOAT                                       NOT NULL,
     date   VARCHAR(64)                                 NOT NULL
 );
